backend/requirements.txt
# NEXORA Backend Requirements
# =============================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0
httpx==0.26.0

# Environment & Configuration
python-dotenv==1.0.0

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.25

# AI & ML
openai==1.10.0
huggingface-hub==0.20.3

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# PDF Generation for Idea Validation Agent
reportlab==4.0.7

# DOCX Generation for Business Planning Agent
python-docx==1.1.0

# PPTX Generation for Pitch Deck Agent
python-pptx==0.6.23

# MVP Builder Dependencies
aiohttp==3.9.1  # Already included above
requests==2.31.0  # Already included above

# Quick Fixes
aiofiles==23.2.1

# Payment Providers
razorpay==1.4.1
stripe==7.10.0  # Already included above

# E2B Sandbox Integration
# e2b==0.15.0  # E2B SDK (if needed for direct integration)

# FireCrawl Integration (using requests/aiohttp)
# firecrawl-py==0.0.8  # Optional: dedicated FireCrawl SDK

# Additional AI Models
# groq==0.4.1  # Optional: dedicated Groq SDK
# anthropic==0.8.1  # Optional: Anthropic SDK

# File Processing
# zipfile and tempfile are built-in Python modules
# Additional file processing utilities can be added here if needed

# Quick Fixes - Security & Performance
slowapi==0.1.9
redis==5.0.1
bleach==6.1.0
sentry-sdk[fastapi]==1.40.0
